requirements.txt
requests~=2.31.0
pyTelegramBotAPI~=4.16.1
python-dateutil~=2.9.0.post0